Added GLWidget. Preparing to make the cut over to QApplication instead of Cinder BasicApp.

+++ src/MapView.h
@@ -22,7 +22,7 @@ public:
         
     void mouseDown(const Vec2i &mousePos);
     
-    void mouseWheel(const float);
+    void mouseWheel(const int);
     
     void mouseDrag(const Vec2i &mousePos, 
                    bool leftDown,
